requirements.txt
numpy
pandas
scipy
Pillow
matplotlib 